separating process file
